Refactor students and TAs in group memberships
