+++ routes/groups.js
@@ -1,7 +1,6 @@
 const router = require('express').Router({ mergeParams: true });
 const groups = require('../controllers/groups_controller');
 const group_memberships_routes = require('./group_memberships');
-const TAs = require('../controllers/TAs_controller');
 const log = require('../helpers/logging');
 const { ensureAuthorised } = require('../permissions');
 const { canShowGroup, canCreateGroup, canUpdateGroup, canDeleteGroup } = require('../permissions/groups');
@@ -41,18 +40,8 @@ router.put(
   ensureAuthorised(req => canUpdateGroup(req.user, req.params.semester_name, req.params.module_code)),
   groups.update
 );
-router.get(
-  '/:name/TAs',
-  ensureAuthorised(req => canUpdateGroup(req.user, req.params.semester_name, req.params.module_code)),
-  TAs.edit
-);
-router.post(
-  '/:name/TAs',
-  ensureAuthorised(req => canUpdateGroup(req.user, req.params.semester_name, req.params.module_code)),
-  TAs.update
-);
 
 // Nest member routes within groups
-router.use('/:name/students', group_memberships_routes);
+router.use('/:name', group_memberships_routes);
 
 module.exports = router;
