--- routes/helpers/TAs.js
@@ -1,9 +1,0 @@
-const groups = require('./groups');
-
-function TAsPath(semester_name, module_code, group_name) {
-  return `${groups.groupPath(semester_name, module_code, group_name)}/TAs`;
-}
-
-module.exports = {
-  TAsPath
-};
