+++ routes/helpers/groups.js
@@ -16,20 +16,9 @@ function groupEditPath(semester_name, module_code, group_name) {
   return `${groupPath(semester_name, module_code, group_name)}/edit`;
 }
 
-// students as group members
-function groupStudentsPath(semester_name, module_code, group_name) {
-  return `${groupPath(semester_name, module_code, group_name)}/students`;
-}
-
-function groupStudentPath(semester_name, module_code, group_name, user_id) {
-  return `${groupStudentsPath(semester_name, module_code, group_name)}/${encodeURIComponent(user_id)}`;
-}
-
 module.exports = {
   groupsPath,
   groupPath,
   groupNewPath,
-  groupEditPath,
-  groupStudentsPath,
-  groupStudentPath
+  groupEditPath
 };
